pyproject.toml
[tool.poetry]
name = "ucan"
version = "0.1.0"
description = "UCAN - Universal Conversational Assistant Navigator"
authors = ["Your Name <your.email@example.com>"]
license = "MIT"
readme = "README.md"
packages = [{ include = "ucan" }]

[tool.poetry.dependencies]
python = ">=3.9,<3.12"
qasync = "^0.27.0"
openai = "^1.1.1"
anthropic = "^0.5.0"
python-dotenv = "^1.0.0"
httpx = "^0.24.1"
requests = "^2.30.0"
watchdog = "^6.0.0"
markdown = "^3.5.0"
duckdb = "^0.10.0"
pyside6 = "^6.8.2.1"
dearpygui = "^1.10.1"
pyyaml = "^6.0.1"
sqlalchemy = "^2.0.27"
alembic = "^1.13.1"
pillow = "^11.1.0"
cairosvg = "^2.7.1"

[tool.poetry.group.dev.dependencies]
black = "^24.2.0"
isort = "^5.13.2"
mypy = "^1.8.0"
pylint = "^3.1.0"
pytest = "^8.0.2"
pytest-qt = "^4.4.0"
pytest-cov = "^4.1.0"
pytest-asyncio = "^0.23.5"

[tool.poetry.scripts]
ucan = "ucan.cli:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
skip = ["docs"]

[tool.black]
line-length = 100
target-version = ['py310']
include = '\.pyi?$'
extend-exclude = '''
# A regex preceded with ^/ will apply only to files and directories
# in the root of the project.
^/docs
'''

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"

[tool.pylint.messages_control]
disable = [
    "C0111", # missing-docstring
    "C0103", # invalid-name
    "C0330", # bad-continuation
    "C0326", # bad-whitespace
    "W0511", # fixme
    "R0903", # too-few-public-methods
]

[tool.pylint.format]
max-line-length = 88
